sport all

+++ src/main/java/it/contrader/dto/ParametersDTO.java
@@ -5,8 +5,8 @@ public class ParametersDTO {
 
 	private int idParameter;
 	private String sesso;
-	private double altezza;
-	private double peso;
+	private int altezza;
+	private int peso;
 	private  int eta;
 	private int polso;
 	private int freqResp;
@@ -16,7 +16,7 @@ public class ParametersDTO {
 		super();
 	}
 
-	public ParametersDTO( String sesso, double altezza, double peso, int eta, int polso, int freqResp,
+	public ParametersDTO( String sesso, int altezza, int peso, int eta, int polso, int freqResp,
 			int tempCorp) {
 		super();
 		
@@ -45,19 +45,19 @@ public class ParametersDTO {
 		this.sesso = sesso;
 	}
 
-	public double getAltezza() {
+	public int getAltezza() {
 		return altezza;
 	}
 
-	public void setAltezza(double altezza) {
+	public void setAltezza(int altezza) {
 		this.altezza = altezza;
 	}
 
-	public double getPeso() {
+	public int getPeso() {
 		return peso;
 	}
 
-	public void setPeso(double peso) {
+	public void setPeso(int peso) {
 		this.peso = peso;
 	}
